nss: change initialization back to read-only
This reverts commit ef9fb4220947844b0d8c619f563427bd274d778c.

As it turns out NSS doesn't force us to use permanent keys and
certificates. So we can switch back to read-only mode without any
database and keep all the stuff in memory.

As the TLS-DSK certificates only have a validity time of a day, it
doesn't make sense to store them anyway and instead request a new one
for each reauthentication.