tls: add new module for TLS handshake messages
I looked at the TLS APIs of GnuTLS/NSS/OpenSSL. None of them offers an
interface to the internals of TLS, i.e. so that we can trigger the
generation of certain messages and then get those in raw binary format.
Their APIs are all geared toward code that wants to layer TLS ontop of
an existing communication channel.

TLS-DSK requires the TLS handshake messages for the GSSAPI data in the
authentication phase. All the above APIs offer is "start handshake and
return once it is complete".

So we will have to roll our own limited TLS implementation...