digest: save context in sipe_digest_md5/sha1_update()
TLS needs to calculate the hashes multiple time but finalizing the PK11
context destroys the collected information. Store the context before
calculating the digest and restore it afterwards.

Now TLS calculates the correct MACs, i.e. the tester can now
successfully complete a handshake with the OpenSSL server. Also the
plugin now goes through to the final TLS-DSK handshake with the Lync
server, but as we have a hard-coded limit of 3 REGISTER retries we abort
too early to succeed.

Disabled crypto secrets debugging in sipe-tls.c again.