sign: fix protocol name for TLS-DSK
sipmsg_breakdown_parse() was hard-coded to Kerberos or NTLM. Now we
supply the protocol name directly from sip-transport.c and generate the
correct message signature for TLS-DSK.

This was the last piece missing to get a successful MAC calculation and
verification with TLS-DSK.

                 *** THE EAGLE HAS LANDED!!! ***

BTW: Lync365 rejects non-Lync clients with a non-standard 403 Forbidden
response. Make sure to change the User Agent value in your account!