purple: change authentication scheme to option list
SIPE currently only supports one active authentication scheme at a time.
With the addition of TLS-DSK support we now have 1 or 2 checkboxes in
the Purple Account UI. It is not clear for the user that they are
mutually exclusive and TLS-DSK has the higher priority.

Replaced the checkboxes with an option list. Users of Kerberos or
TLS-DSK will have to update their account settings!

SIPE support in telepathy-haze might need update too...