sipe: improve error handling for contact search
As reported in the following forum entry:

<https://sourceforge.net/projects/sipe/forums/forum/688534/topic/4720222>

there are some installations that have disabled the old(?) directory
search. It could be that this is the default setting for Lync or at
least Lync/Office365, where I have encountered it.

process_search_contact_response() now shows an error dialog to the user
for the following cases:

  - the response code is not 200 OK
  - the XML content in the body is corrupted
  - no contacts found. IMHO that is a better response than an empty
    result window.

Hopefully some logs from the M$ Lync client will reveal how the contact
search works there.