soap: factor out SOAP over SIP functions
Moved code from sipe.c to new module sip-soap.c. Note that the naming
convention is the same as for sip-transport.c, because this is only for
SOAP message transferred via SIP SERVICE messages.

Refactored the functionality so that the common SOAP stuff is in
sip-soap.c and not in some preprocessor macro.