core cleanup: move [MS-SIP] deltanum counters to sipe_private
This makes sip-soap.c sip_account_data-free and removes the need to
include sipe.h.

I noticed that sipe.c was using g_ascii_strtod() to convert integers.
I'm not sure if that has ever worked correctly, because that function
returns a double, not an integer.

Fixed the naming in sip-soap.c. Those SOAP messages are from [MS-SIP]
and not [MS-PRES].