certificate: switch from Anon to WinNegotiate port
The WebTicketServiceAnon authentication port doesn't seem to work at
all. Use the WebTicketServiceWinNegotiate port instead. For this URI the
authentication against the WebTicket service is not handled inside SOAP
using WS-Trust elements but on the underlying HTTP level.

The HTTP authentication is only accepted for the SOAP action it was
requested for. As this happens on the HTTP level we need to add a
SOAPAction: header. Updated http_conn API to be able to provide
additional HTTP headers.

With these changes it is possible to authenticate against internal Lync
installations. At least against the one I got a test account for :-)