core cleanup: start to separate buddy menu code
As this change is too big to do in one go I have chosen to take a
piecemeal approach.

This breaks the compilation of sipe.c, so the first step was to remove
sipe.c from the build. The module is still included in the tarball,
because otherwise the intltool check will fail.

Removed sipe.h and core-depurple.h and updated sipe-core.c and
purple-plugin.c accordingly.

Removed all purple stuff from src/core/Makefile.am. I also noticed that
we still compiled sipe-utils.c as purple module under Windows, although
that purple dependency had already been removed months ago.

All of the above means that finally the core directory is purple-free!

As an initial step for the buddy menu cleanup the code for the first 4
entries, all related to chat, was separated and moved to the appropriate
modules in core and purple directories.