sspi: use "Schannel" as TLS-DSK provider
This is a much better educated guess then the previous choice, because
Schannel is the SSL/TLS provider for Windows programs. It still doesn't
work, but at least we no longer fail with

  AcquireCredentialsHandleA: The requested security package does not exist

To get this working we will most likely have to figure out how to pass
the client certificate, which is stored in NSS binary format, down to
the SSPI layer in a form that Schannel will accept.