svc: add HTTP session handling
Microsoft seems to have enabled server load balancing on Office365.
Compared to the logs taken at the end of last year, the new logs now
show that the server responses contain HTTP Set-Cookie: headers. We need
to store the cookie and send it in all related SOAP requests so that the
same server processes them.

The highest layer, i.e. Certificate or MS-DLX, must now allocate a
session which will be re-used in all requests.