tls: fix incorrect length for certificate message
tls_client_certificate() didn't include the 3 bytes for the outer vector
length in the message size when calling compile_handshake_msg(). This
caused a buffer overrun when the code was compiled on a 32-bit platform.

This must be the problem commit b2418af604cc274b08351aa632b264e1dec6d95f
was intended to work around.