telepathy: implement sipe_backend_network_ip_address
Changed this backend API to be more libpurple independent by requiring
sipe_core_public as parameter. Updated core/miranda/purple code
accordingly.

In telepathy we retrieve the local IP address from the server transport
socket, if we already have a connection and cache this string for all
later calls. If the retrieval should fail for any reason, we fall back
to 127.0.0.1.