telepathy: fix read of large messages in transport
If a message is larger than the currently allocated buffer, then the
input routine in the core will NOT consume any data. Thus we're setting
up the next read for 0 bytes which incorrectly lead to a server
disconnected situation.

If the next read would be for 0 bytes we increase the buffer before
kicking off the next read.

The code now completes a complete login scenario with the OCS server.