telepathy: always check cancel in transport async ops
Successful completion of an asynchronous operation might already be
queued in the main event loop when disconnect is called on the
transport. That means that just checking for an error return is not
enough to ensure that the transport is still valid.

Also reordered freeing up of the transport data structure to make sure
we are not leaking the cancel object.