telepathy: fix crash from read_completed during flush
When a transport has the do_flush flag set then none of the outstanding
async operations gets cancelled, because they all share the same cancel
object. So it can happen that read_completed is called when the SIPE
core has already been deallocated.

Simply ignore the read result in this case.