telepathy: make contact search channel work
Found the reason why our new channel wasn't recognized. Now the "Find"
button in Empathy gets activated. But it ignores our available search
fields and simply sends us the contents of the search box with the label
"fn" (full name).

I wonder what the whole "available search keys" is for, if the UI simply
ignores it? Is the search channel supposed to parse "fn" itself?