core: improve core/backend search API
Telepathy can start multiple searches and needs to be able to identify
which search results belong to which search channel.

Introduce an opaque token that the backend can supply when it requests a
search from the core. This token will be presented by the core when it
either tells the backend that search results are available or the search
has failed. Purple and Miranda backends ignore this token and simply
supply NULL.

Refactored search failure paths to call a new backend function with the
token and the message. For purple and Miranda backends this function is
simply implemented by calling sipe_backend_notify_error() like before.

Added search failure path for empty/invalid query.

Refactored SOAP code in sipe-buddy.c to use a common function. This new
function search_soap_request() implements the transaction payload that
carries the token to the callback.

Updated transactions_remove() to support transaction payloads without
destroy notifier.