webticket: add detection for ADFS setup
If the RealmInfo XML data contains a STSAuthURL node then this means
that the account is from a federated domain, i.e. there exists an ADFS
server for the AD to generate Web Ticket tokens. ADFS support is
currently not implemented, so give feedback to the user.

This implmentation is based on observed behavior. It is unkown if it is
officially documented somewhere.

It is currently unknown if it will be possible to implement ADFS
support in SIPE.