tls-dsk: extract expires value from TLS certificate
[MS-SIPAE] Section 3.2.2 states that the expiration time for the TLS-DSK
SA must be the lesser of the certificate expiration time or 8 hours. Add
the necessary interfaces to sipe-tls and sipe-cert-crypto to extract the
expiration time from the certificate.

Clean up the reauthentication timeout in sip-transport.c. Either the
authentication scheme provides a valid expiration time or not. If it is
valid, i.e. longer than 5 minutes, then use it, otherwise use a default
of 8 hours.

Triggered by analysis for bug #3580212.