webticket: implement queueing
When a Web Ticket request is pending for a Service URI and another
request is started for the same URI, then we simply queue the new
request on the already pending one. Once the pending one is completed it
will call its callback and all callbacks from the queued list. This will
speed up all operations that start multiple requests in a row.