Feature #3578132: Kerberos configuration should be passwordless
Introduce a new core API that the backend can check before calling
sipe_core_allocate() to check if the selected authentication scheme
requires a password or not. Update the sip-sec mechanisms to provide a
new hook function for this purpose.

Now we can use OPT_PROTO_PASSWORD_OPTIONAL in the purple backend and
inform the user when he needs to enter a password.