Feature #3578132: Kerberos configuration should be passwordless (II)
There are Pidgin users out there that don't like that their password
has now to be stored in cleartext on the disk. For them the original
implementation is not acceptable. Refactor the functionality of
sipe_purple_login() and use purple_account_request_password() when the
user hasn't set a password.

Sometimes reading the documentation *FIRST* would help...

Re-arranging the code also revealed that the original implementation
probably caused the SSO flag setting to be ignored on Windows SSPI
configurations.