security: fix libkrb5 flagging for Windows
Restructured hook function #define's in sip-sec.c to be per scheme
instead of per platform. This is much simpler and easier to understand
then the previous approach.

Now you can compile in libkrb5 on Windows platform if you don't want to
rely on SSPI. Makefile.mingw doesn't have support for this, though...