http: implement 3xx response handling
Redirection requires URL parsing. Make the URL parsing visible to the
request layer. As the data structure is now visible we can simplify the
new request API by passing the parsed URI data structure.

With this change the new implementation should be on par with the old
implementation in http-conn.c.