Fix #190: SIP 407 response rejected with invalid message signature
NOTE: SIPE project on SourceForge got updated to the new project format and
      therefore all bug & feature request numbers have changed.

The log for this issue shows a 407 response with the follwing headers:

  Authentication-Info: NTLM rspauth="...
  Proxy-Authenticate: Digest realm="...

The first is the one we need to verify the signature of the message.
The second one is needed to process the 407 response.

But sipmsg_breakdown_parse() incorrectly was looking for both lines with
"Proxy-Authenticate" having the higher priority. Therefore the wrong data
was fed to the signature calculation, causing the failure and subsequently
the connection drop.

Drop "Proxy-Authenticate" from the parser to fix this.