transport: reset subscribed flag for reauthentication
This is a temporary fix for an issue noticed in the logs provided for
bug #186: SIPE doesn't honor the expiration value for event
subscriptions. After the event has expired the server will no longer
send us BENOTIFY messages with the data and SIPE gets out of sync with
the server.

F.ex. the log shows that after "vnd-microsoft-roaming-self" expires,
SIPE can no longer update the "version" field for the categories. When
SIPE tries to send a "own presence" update, it will be rejected by the
server with "409 Conflict". SIPE then updates the categories from the
server response and retries. But the next attempt will fail again with
the same problem.

The logs show that the epxiration value for the event subsriptions are
way longer than the re-authentication timeout. So we simple reset the
"subscribed" flag before initiating a re-authentication. When that is
successfully completed, SIPE will subscribe to the events just like it
did at the initial successful authentication attempt.