security: add 2 test for Digest authentication
The new code (based on the originally removed code) didn't pass the
tests at all, because it was not following the specifications. My guess
is that the whole 407/Proxy Authentication code was never tested and has
never worked.

Added two Digest test samples and fixed the Digest authentication code
to pass those two tests.

I'm not confident that Proxy Authentication will work with this code,
but this is the best we can do at the moment without further logs.