telepathy: certificate accepted path in callback
Now user can accept/reject a certificate. On acceptance the connection
attempt will be retried, but this time we detect that we asked the user
already and tell the TLS handshake to proceed.

We only need to figure out what certificate data telepathy clients, like
Empathy, really expect to be able to show it to the user.