Fix #194: Unable to be heard when answering to external call
When we receive a multipart/alternative INVITE to a call, it is possible
for a part not marked as 'ms-proxy-2007fallback' to entirely omit the
Content-Disposition	header. For such messages sipe_invite_mime_cb()
possibly didn't select a message part at all and therefore whole multipart
body was parsed as SDP, leading to unpredictable results.

The part without C-D header is now picked as the single alternative for
further processing.