Fix #193: Pidgin Status changes stop working
When we subscribe to a new user with batched subscription then the
response will only contain that user. In order not to loose the
information about all the other subscribed users we now merge the old
and the new buddies list in the subscription data structure.

TODO; check if this works correctly also when we remove a buddy.