Fix #193: Pidgin Status changes stop working (III)
The previous commit only worked, because the SUBSCRIBE message was
invalid in another way and generated an error message. It looked that it
worked in testing, though...

Add the missing "Supported: eventlist" header. But now we're back at
square 1, because the server terminates our old subscription again.

After studying [MS-PRES] and [MS-SIP] it seems that we should only send
the SUBSCRIBE message to the added buddy URI for the case when the
server tells us to re-subscribe. Sending the initial single SUBSCRIBE to
self URI seems to fix the termination problem.