http: Fix crash when talking to flaky OCS server.
We tend to see lots of SSL read and write errors on an OCS server we
currently use. Most of these are handled well and the plugin reconnects
seamlessly, however due to some recent infrastructure changes we
started seeing pidgin segfault as often as every couple of minutes.

It appears that if a connection is successfully made, but immediately
sees a read error, it is then dropped but the timeout schedule is not
cancelled.

When the http timeout schedule then fires, the already freed connection
is dropped a second time.