notify: contact list processing for Lync 2013 UCS
When the contact list has been migrated to Unified Contact Store (UCS)
on Lync 2013 then the list piggybacked on the SUBSCRIBE response is
empty. That means that we won't subscribe to any buddies and therefore
they all appear offline.

But the server immediately sends a NOTIFY with the real list. Change the
list processing to ignore the first empty list and process the second
one.