backend: add time_t to sipe_backend_chat_message()
We need to be able to specify a timestamp if we want to replay messages
from history after joining a channel. A timestamp of 0 signifies to use
the current time.

libpurple backend simply forwards the time_t value instead of
hard-coding to time(NULLL).

Miranda backend is missing timestamp handling (see comments).