ucs: retrieve IM list after update trigger
When another client modifies the UCS contact list, then we receive a
NOTIFY with the updated roaming contact list. Use this as a trigger to
request the IM list again from UCS.

Make sure that group & buddy creation functions correctly skip when
trying to trying to create an already existing group/buddy.

If the update adds a new buddy, then we need to subscribe to presence
for it.

NOTE: currently group or buddy removals are not detected, i.e. those
will only correctly represented after reconnecting to the account.