buddy: check for obsolete groups
Before we process the updated UCS contact list we set the obsolete flag
on all groups a buddy is associated with. When a buddy is still
associated with a group on the list then this flag gets reset. After
processing the list we remove the buddy from all the groups which still
have the obsolete flag set.

This handles the case when a buddy is associated with multiple groups
and is removed from some groups by another client.