group: add interfaces for obsolete group cleanup
Before we process the updated list we set the obsolete flag on all
groups. When a group is still on the list then this flag gets reset.
After processing the list we remove all groups which still have the
obsolete flag set.

This should complete the support for UCS contact list updates.