ucs: optimize contact list update requests
The server sends us a NOTIFY with an update contact list after we have
made changes to UCS. Instead of generating a useless GetImListItems
operation we now ignore those update triggers that arrive within 10
seconds of our last response.

Not perfect, but this reduces the network traffic considerably when the
user is making many changes to the buddy list in a short time.