security: only throw away completed contexts
In sip_sec_init_sec_context__krb5() we threw away an already existing
Kerberos context when a new authentication handshake was started on the
existing connection. Although for Kerberos this is not a really a
problem, because it has only one authentication step, this is not
correct for the general GSSAPI case.

We now check if SIP_SEC_FLAG_COMMON_READY is set.