Fix #222: SIPE crashes when groupchat session expires (I)
The debug log shows that for some unknown reason the groupchat session
has expired and the server responds with 481. This response causes SIPE
to automatically close the dialog. The next time we send try to send a
message to the groupchat session we get NULL for a dialog. We didn't
check for this condition and therefore the code crashes.

As a quick fix we now check dialog == NULL and don't send the message.

(cherry picked from commit eeb27a3e53c4591f417b6d8ce3cbb738a36a4de5)