ucs: fix Persona key extraction
Based on the log provided in the following discussion thread

  https://sourceforge.net/p/sipe/discussion/688534/thread/db025884/

it turned out that the text for Attribution/DisplayName is localized,
i.e. the text "Lync Contacts" gets translated to the local language.

Switch over to check IsQuickContact == true and IsHidden == false, which
seems to be another indication that the Attribution node contains the
key for the Lync Contact. It's still a guessing game :-(

(cherry picked from commit 8763852bb5aa9709bbe6f93b9088f8d607e061b8)