adium: cleanup configurePurpleAccount
We control KEY_CONNECT_HOST and the accounts.xml "server" fields directly.  Easiest way to control autodiscover is simply to set everything the way want it and skip call ing the superclass configurePurpleAccount.  It doesnt set anything we need anyway.

Likewise, no need to hook into reachability monitor if we simply set our "dummy" KEY_CONNECT_HOST to something that should always be reachable:  localhost

Implement another hash and condense the function into a for-loop and an if-else loop