security: set GSS_C_MUTUAL_FLAG only when required
We should not have to set GSS_C_MUTUAL_FLAG when we don't have to. But
unfortunately older Microsoft SPNEGO implementations send a final empty
packet that is not accepted by older MIT KRB5 implementations. Switch to
a runtime decision if we need the GSS_C_MUTUAL_FLAG workaround or not.

Added comments explaining the details for this workaround.