purple: ignore empty search values
It seems that Adium is not 100% compatible to Pidgin/libpurple. When the
contact search callback is called, then

   purple_request_field_string_get_value()

does not return NULL for empty string fields, but an empty string. Empty
strings in the XML request are rejected by the server as illegal, i.e.
contact search doesn't work at all in Adium.

Add a check that the string length is greater than 0.