Fix #238: False "not delivered" in conference
When Lync Archiving option is enabled, then the IMDN (Instant Message
Delivery Notification) XML for conference messages contains "recipient"
nodes also in the case of successful delivery, with a status code of 200.

sipe_process_imdn() now parses the "status" node and only generates a
"message not delivered" error if that node is missing or the error code
is greater or equal to 300.