Fix #210: Conference call ends with error message (II)
Looks like we don't have a reliable way to recognize which ICE dialect
to use to communicate with conferencing server (perhaps it's hard-coded
in Lync/OCS client corresponding to the server release?). Lacking a
better option let's just retry the call with the other ICE version when
the first attempt fails.