ews: add 4th autodiscover type
The 3rd type introduced to fix problems with Office 365 in

    commit 0ba825ca063e3d94487ec518e9a1c3f5a722dd1c

was based on trial & error. It seems that sending a POST request to a
http:// autodiscover URL does not work for all accounts:

    https://sourceforge.net/p/sipe/discussion/688534/thread/d21ebf11/

After studying another M$ document it turns out that we are allowed to
also send a GET request to such an URL and will then receive a redirect
response.

I tested that this works for at least 2 different Office 365 accounts.
GET seems to be accepted for Office 365 accounts for which POST was also
working. So we now try GET first and only then we try POST.