Fix #248: Remove libpurple SSL configure check
On some installations the stripped-down libpurple initialization is not
enough and causes the SSL configure check to fail. A better solution
would be to call purple_core_init(), but that fails on other platforms.

The only option is to remove the configure check completely. Not a big
loss anyway...

(cherry picked from commit f79a5d68069ac52c24138abdf4f4724cb88c245a)