Fix #255: Crash when PersistentChat sends BYE instead of response
The log provided by the user shows that he is probably not allowed to
use PersistentChat. After a successful INVITE our first XCCOS command
causes the server to send a BYE, which causes SIPE to close the dialog.
Our 200 response to the BYE triggers a 481 response to which SIPE tries
to answer with a BYE without checking that the dialog is already closed.

Fix this by suppressing the BYE when dialog is NULL.