Fix #259: HTML response to EWS autodiscover triggers libxml2 assert (II)
It seems that some HTTP servers ignore the "Accept:" header. Added a
check that the "Content-Type:" header starts with "text/xml" before we
call the libxml2 parser.

NOTE: I couldn't reproduce the libxml2 assert on Fedora 20 from the
provided debug log. It simply complains about XML errors and returns
a parse error to the caller.