Fix #258: V&V call gets rejected when IPv6 is enabled
commit 1556f377ecc7c3d4887f8e9f2805a4e7f7e60660 changed the candidates
list to be sorted when it is created and *NOT* sort it when it is
converted to the SDP message. Unfortunately that changed the value of
msg->ip in sipe_media_to_sdpmsg() [through backend_stream_to_sdpmedia()
and backend_candidates_to_sdpcandidate()] in the case when an IPv6
address was assigned to the host.

As the media code does not support IPv6 addresses yet, we now reject IP
address strings with ":" in get_stream_ip_and_ports().

(cherry picked from commit 396613000a879625c5a836a3b5f71b8f3e20fd0a)