Fix #262: Adium: SIPE doesn't auto- or re-connect
If server auto-discovery is in use then the plugin doesn't have a value
for KEY_CONNECT_HOST. We used to set it to the localized name returned
by NSHost. But that isn't necessarily a valid DNS hostname and therefore
Adium's reachability test fails for the SIPE account.

Now we hard-code the value to "localhost" instead. During my tests this
was the only value that always worked.