purple: don't ignore status change when idle-away
commit 74820ce888af76c76c89e05d229aece0de33f443 added a check to ignore
state changes when the current libpurple state is idle or away. It is
not clear why this check was added. Simply remove it.

This should fix the long-standing issue that Pidgin status was not
returning to available after a meeting ended.

(cherry picked from commit c58bb402552a3fd2e7a083ecae730f600dde57a5)