svc: use authuser for RealInfo request
A user reported in this forum entry

  https://sourceforge.net/p/sipe/discussion/688534/thread/e7d7035e/#d3b1

that he gets a different response to the RealmInfo request if it uses
username or authuser. Only the latter response allows him to login
successfully.

Use authuser for the RealmInfo request, but only if it looks like
"user@domain". I.e. authuser will not be used if the user has specified
DOMAIN\account as login name.