ucs: first stab at search functionality
If we are on Lync 2013 and contacts have been migrated to UCS then
sipe_core_buddy_search() forwards the search request to
sipe_ucs_search(). We use the EWS FindPeople operation to search the
GAL.

In this first attempt we use given_name as query string.

There seem to be some restrictions:

- "m:Restriction" doesn't accept "persona:ImAddress". So our search
  results contain Personas without ImAddress node which we then need to
  filter out by hand.
- There seems to be no field corresponding to "country". The one that
  comes closest "persona:Locations" doesn't seem to work.