Fix #278: 488 error after libnice upgrade
When libnice 0.1.10 is combined with Farstream 0.1.x, TCP candidates
that libnice now always generates are incorrectly labeled as UDP by
Farstream (which has no support for TCP in 0.1 branch). We try to filter
them in SIPE, assuming that RTP+RTCP UDP pairs that share the same port
are actually mistagged TCP candidates.