cal: improve periodic calendar update
The default granularity in Exchange is 15 minutes, so it makes sense to
check for calendar updates every 15 minutes, not 30 minutes.

Instead of a fixed update interval, schedule the update 30 seconds
before the next calendar interval starts. This will make sure that OCS
has been updated with the latest information before the next calendar
interval starts.