transport: SIP SA always expires after 8 hours
Triggered by the discussion of an endless 401 loop in

  https://sourceforge.net/p/sipe/discussion/688534/thread/1a259300/

The "expires" value for SIP was handled inconsistently:

 - NTLM:     8 hours
 - Kerberos: ticket expiry time, can be higher than 8 hours
 - TLS-DSK:  certificate expiration time, max 8 hours

But [MS-SIPAE] 3.2.2 "Timers" clearly states that the maximum validity
of the SIP SA is 8 hours. Remove all special expires handling from other
places and handle it only in process_register_response().

Also schedule the re-register event 30 seconds before the timeout. This
is more in line with what we do for other timeouts already.