ucs: notify user about initialization failure
EWS is required for UCS to work. Currently failures are only reported to
the debug log. This has let to multiple forum posts and unnecessary bug
reports.

Add a user notification so that the user knows that either the default
settings do not work or his Email settings are incorrect.