srtp: honor encryption key ID in INVITE response
Media encryption key we offer in a response to a call invitation should
have the same ID as that 'a=crypto' line from the caller's INVITE which
we want to use for encrypting the data stream.

So far the ID was hardcoded as '1', which was causing trouble
communicating with some endpoints, namely when we were accepting a call
from the public telephone network through a PSTN gateway.