media: don't poll backend for established candidate pairs
Rework the code to better utilize "candidate-pair-established" signal
from PurpleMedia. With this change there's no need for repeated polling
of active candidate pairs in reinvite_on_candidate_pair_cb(), and by
removing that callback we also avoid a segfault when the outgoing call
was cancelled before establishing the candidate pair.