appshare: fix random conference connect failure
Before we can launch an RDP client and initiate RDP communication, two
prerequisities have to be fulfilled:

1. connection is established and the socket writable
2. SDP media negotiation is concluded on SIP layer

We didn't check the second condition and so it used to happen very often
that we launched xfreerdp before we received the last INVITE response.
The first packet of RDP communication was sent, but we never got any
reply because our counterparty RDP server wasn't listening yet.