conf: advertise mute state to the Lync server
During a group call it appears as if the users of pidgin clients were
talking all the time, even when they are actually on mute.

When everyone else from the call is muted, it is distracting to Skype
for Business users from the same call, who may think that the Sipe user
has something to say.

The mute state should be properly advertised to the Lync server.