conf: remove Lync URL parsing code
M$ has changed the Lync meeting HTML pages in Skype for Business. They
now are dynamic pages which may or may not generate a conference SIP
URI. As our HTML parser can no longer find a SIP URI the code falls back
to the old URL parser. That can produce an incorrect result which means
joining a conference can fail for a user with a meaningless DNS error
message.

Remove the URL parser and replace it with an error message.