media: send MS-RTP video source request
We first have to wait for FsSession belonging to our sipe_media_stream
to appear and connect "on-sending-rtcp" signal of its internal RTP
session. Then we can tack VSR to the regular batch of RTCP messages in
the signal handler.

This change adds a dependency on two pkg-config packages when building
purple backend with VV: gstreamer-rtp-(1.0|0.10) and farstream-0.2.