Fail to configure if --with-vv specified and VV not available
Those build breakages I just fixed... were actually useful for me
because I didn't *want* it to build without voice and video support
anyway. We do need to fix them, but let's do this too so that we can
get the "VV or bust" that we actually want in the distribution
packages, and never a silent lack of the feature.

Note that I haven't tested with VV enabled but not FreeRDP (in which
case I believe screen-share RX ought to work, but not TX). And we
know that the telepathy build is hosed too.