purple: don't pre-fill new buddy alias from search result
User has an option to fill in an alias, but by default we should keep
the value empty so that the regular display name from contact card is
used.

Contacts added from MS clients always have empty "name" attribute
in the contact list - it appears at least the latest versions don't
even support assigning an alias to a contact.