media: multichannel needs GStreamer 1.6
Lync offers multichannel audio as a codec with the same encoding name
as the mono variant, but a different payload type and an extra encoding
parameter:

  a=rtpmap:117 G722/8000/2
  a=rtpmap:9 G722/8000

This causes trouble with GStreamer RTP payloaders prior 1.6 that
supported only default payload types (9 in the case of G722). Let's
ignore the multichannel codecs when used GStreamer version isn't recent
enough so as to avoid not-negotiated caps errors when they are chosen
for transmission.